fix segfaults on OpenBSD: int cannot hold all the values a ptrdiff_t can take.

+++ src/libregexp/regcomp.c
@@ -238,7 +238,7 @@ optimize(Reprog *pp)
 	int size;
 	Reprog *npp;
 	Reclass *cl;
-	int diff;
+	ptrdiff_t diff;
 
 	/*
 	 *  get rid of NOOP chains
